requirements.txt
flask==3.0.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.1
joblib==1.4.2
imbalanced-learn==0.12.3
openai>=1.0.0
python-dotenv==1.0.1